+++ doc/release_notes.rst
@@ -4,6 +4,7 @@ Release Notes
 Upcoming Version
 ----------------
 
+* The writing and reading from netcdf files was fixed to correctly handle the model `parameters` field.
 
 Version 0.3.10
 --------------